Role 추가 및 테이블 구조 변경 유저-Role-권한 관계 구조 변경

+++ backend/src/main/java/com/datamindhub/blog/repository/AuthorityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorityRepository extends JpaRepository<Authority, Byte> {
+public interface AuthorityRepository extends JpaRepository<Authority, Integer> {
     Authority findByName(String name);
 }
